flaskapp/requirements.txt
git+http://github.com/mitsuhiko/flask.git

flask-classy
unittest